+++ src/pages/Admin/components/Offer/OfferDetailContent/index.ts
