+++ src/pages/Admin/components/Offer/OfferDetailInfo/MainTitle/container.tsx
@@ -1,6 +1,6 @@
 import React, { FunctionComponent } from "react";
 
-import { MainTitle } from "../../../../../../components/MainTitle";
+import { MainTitle } from "../../../MainTitle";
 import { useTranslations } from "$hooks/queries";
 import { IOffer } from "$interfaces/Offer";
