+++ src/pages/Admin/components/Offer/OfferDetailInfo/MainTitle/index.ts
