+++ src/pages/Admin/components/Offer/OfferDetailInfo/OfferDetails/component.tsx
@@ -1,5 +1,5 @@
 import React, { FunctionComponent } from "react";
-import { TaskHeaderInfo } from "../../../../../../components/TaskHeaderInfo";
+import { TaskHeaderInfo } from "../../../TaskHeaderInfo";
 import styles from "./styles.module.scss";
 import { NumberFormatter } from "$models/NumberFormatter";
