+++ src/pages/Admin/components/Offer/OfferDetailInfo/OfferDetails/container.tsx
