+++ src/pages/Admin/components/Offer/OfferDetailInfo/OfferDetails/index.ts
