+++ src/pages/Admin/components/Offer/OfferDetailInfo/OfferDetails/styles.module.scss
