+++ src/pages/Admin/components/Offer/OfferDetailInfo/index.ts
